// FrontEnd/src/app/models/grafico-coluna/grafico-coluna.ts
export class GraficoColunaModel {

    CodMarca: number;
    DescMarca: string;
    BaseAbs: number;
    PercTOM: number;
    TesteSIGTOM: string;
    PercOM: number;
    TesteSIGOM: string;
    PercPrompeted: number;
    TesteSIGPrompeted: string;
    PercTotal: number;
    TesteSigTotal: string;
    PercOutros : number;
    TesteSigOutros : string;
    BaseMinima:string;
}